requirements.txt
fastapi
uvicorn[standard]
pandas
python-docx
readability-lxml
lxml
requests
trafilatura
python-dateutil
